Website/django_backend/requirements.txt
# python3 -m venv myvnv 
# source myvnv/bin/activate
# python3 -m pip install --upgrade pip
# mkdir -p media/private_messages 

asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
bcrypt==4.1.3
certifi==2024.6.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.7
daphne==4.1.2
Django==4.2.13
django-allauth==0.63.3
django-cors-headers==4.3.1
django-password-validators==1.7.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
msgpack==1.0.8
pillow==10.3.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
pyotp==2.9.0
python-dotenv==1.0.1
qrcode==7.0
redis==5.0.4
requests==2.32.3
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.11.0
urllib3==2.2.1
zope.interface==6.3
